+++ projects/_01_ingest/ascwds/tests/fargate/utils/test_clean_workplace_utils.py
@@ -1,4 +1,3 @@
-import unittest
 from datetime import date
 
 import polars as pl